+++ backend/gateway/src/main/java/com/example/gateway/pojo/TokenData.java
@@ -6,7 +6,7 @@ import com.google.gson.JsonSyntaxException;
 
 public class TokenData {
 
-    public Integer userId;
+    public Long userId;
 
     public String username;
