+++ backend/user/src/main/java/com/example/user/entity/UserRole.java
@@ -1,12 +1,13 @@
 package com.example.user.entity;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import java.util.Date;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.io.Serializable;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
+
+import java.io.Serializable;
+import java.util.Date;
 
 /**
  * 用户-角色
@@ -23,7 +24,7 @@ public class UserRole implements Serializable {
 
     private Date createTime;
 
-    private Integer userId;
+    private Long userId;
 
     private Integer roleId;
