+++ backend/user/src/main/java/com/example/user/service/UserRoleService.java
@@ -3,11 +3,12 @@ package com.example.user.service;
 import com.baomidou.mybatisplus.extension.service.IService;
 import com.example.user.entity.UserRole;
 
-import java.util.*;
+import java.util.List;
+import java.util.Map;
 
 public interface UserRoleService extends IService<UserRole> {
 
-    void addUserRole(Integer userId, List<Integer> roleIdList, boolean deleteOldRole);
+    void addUserRole(Long userId, List<Integer> roleIdList, boolean deleteOldRole);
 
     void batchAddUserRole(List<Map<String, Object>> userRoleList);
