+++ backend/user/src/main/java/com/example/user/service/UserService.java
@@ -2,17 +2,16 @@ package com.example.user.service;
 
 import com.baomidou.mybatisplus.extension.service.IService;
 import com.example.common.exceptions.CustomRuntimeException;
-import com.example.user.entity.User;
 import com.example.user.dto.UserDTO;
+import com.example.user.entity.User;
 import org.springframework.data.domain.Page;
 
-import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
 public interface UserService extends IService<User> {
 
-    List<Map<String, Object>> getUserRoleAndPermissionsByUserId(List<Integer> userIds);
+    List<Map<String, Object>> getUserRoleAndPermissionsByUserId(List<Long> userIds);
 
     Page<User> getUserList(String userName, String minCreateTime, String maxCreateTime, String orderBy, String orderMethod, Integer page, Integer pageSize);
